Create AddonCatalogCacheCreator classes and tests
The AddonCatalogCacheCreator.py script is intended to be run server-side
by a cron-like process, regularly generating a downloadable cache of all
addon metadata. The Addon Manager will then fetch this file, which in
conjunction with the Addon Catalog hosted at
https://github.com/FreeCAD/FreeCAD-Addons/AddonCatalog.json
will be enough information to display the list of non-macro addons,
including their icons. As of this writing the cache file is approximately
4MB uncompressed.