Activate the Addon Catalog code
Remove all local cache code except the caching of the raw Addon Catalog JSON, and replace it with a fetch of the catalog's SHA256 hash and determination of whether the cache is stale. Temporarily removes any macro cache whatsoever, which will shortly be replaced by a macro equivalent of the addon cache.